Fixes were added

+++ src/controllers/user.controller.js
@@ -7,16 +7,28 @@ const userService = require('../services/users.service');
 const get = (req, res) => {
   const allUsers = userService.getAllUsers();
 
+  if (!allUsers) {
+    res.status(404).send('Not Found: The specified entity does not exist');
+
+    return;
+  }
+
   res.send(allUsers);
 };
 
 const getOne = (req, res) => {
   const { id } = req.params;
 
+  if (!id) {
+    res.status(400).send('Bad Request: Missing required parameter');
+
+    return;
+  }
+
   const user = userService.getUsersById(id);
 
   if (!user) {
-    res.sendStatus(404);
+    res.status(404).send('Not Found: The specified entity does not exist');
 
     return;
   }
@@ -28,12 +40,18 @@ const create = (req, res) => {
   const { name } = req.body;
 
   if (!name) {
-    res.sendStatus(400);
+    res.status(400).send('Bad Request: Missing required parameter');
 
     return;
-  };
+  }
 
   const user = userService.createUser(name);
+
+  if (!user) {
+    res.status(404).send('Not Found: The specified entity does not exist');
+
+    return;
+  }
 
   res.statusCode = 201;
   res.send(user);
@@ -43,10 +61,18 @@ const update = (req, res) => {
   const { id } = req.params;
   const { name } = req.body;
 
+  if (!name || !id) {
+    res.status(400).send('Bad Request: Missing required parameter');
+
+    return;
+  }
+
   const user = userService.getUsersById(id);
 
   if (!user) {
-    return res.sendStatus(404);
+    res.status(404).send('Not Found: The specified entity does not exist');
+
+    return;
   }
 
   if (typeof name !== 'string') {
@@ -63,10 +89,18 @@ const update = (req, res) => {
 const remove = async(req, res) => {
   const { id } = req.params;
 
+  if (!id) {
+    res.status(400).send('Bad Request: Missing required parameter');
+
+    return;
+  }
+
   const user = userService.getUsersById(id);
 
   if (!user) {
-    return res.sendStatus(404);
+    res.status(404).send('Not Found: The specified entity does not exist');
+
+    return;
   }
 
   userService.removeUser(id);
